+++ src/main/java/com/example/maktabproject1/entity/SubServiceEntity.java
@@ -3,11 +3,12 @@ package com.example.maktabproject1.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "sub_services")
 @Data
 public class SubServiceEntity {
-
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -19,12 +20,9 @@ public class SubServiceEntity {
     private String description;
 
     @Column(nullable = false)
-    private double basePrice;
+    private BigDecimal basePrice;
 
     @ManyToOne
     @JoinColumn(name = "service_category_id", nullable = false)
     private ServiceCategoryEntity serviceCategory;
-
-
 }
-
